requirements.txt
# NL2SQL_AGENT - Requirements
# Python 3.9+ required

# =============================================================================
# Core LangChain & LangGraph
# =============================================================================
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-ollama>=0.0.1
langgraph>=0.0.20

# =============================================================================
# Embeddings & Vector Store
# =============================================================================
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
chromadb>=0.4.0

# =============================================================================
# LLM Providers
# =============================================================================
openai>=1.0.0
# ollama  # Install separately: https://ollama.com/

# =============================================================================
# Database
# =============================================================================
# sqlite3 is built-in to Python

# =============================================================================
# Web Interface
# =============================================================================
streamlit>=1.28.0

# =============================================================================
# Evaluation
# =============================================================================
func-timeout>=4.3.5
numpy>=1.24.0
tqdm>=4.65.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
pandas>=2.0.0

# =============================================================================
# Optional: LangSmith (for tracing)
# =============================================================================
# langsmith>=0.0.50
